[Refactor] GlassBorderAttributes Utils로 이동 (#149)
